+++ noodles-vcf/src/header/info.rs
@@ -1,6 +1,5 @@
 //! VCF header info record components.
 
 pub mod key;
-pub mod ty;
 
-pub use self::{key::Key, ty::Type};
+pub use self::key::Key;