+++ noodles-vcf/CHANGELOG.md
@@ -10,6 +10,10 @@
 
     Use `header::record::value::map::format::Type` instead.
 
+  * vcf/header: Move `header::info::Type` to record map value.
+
+    Use `header::record::value::map::info::Type` instead.
+
 ## 0.24.0 - 2023-02-03
 
 ### Added